move forecast to forecast.py
