+++ nowcasting_datamodel/fake.py
@@ -6,13 +6,12 @@ import numpy as np
 from sqlalchemy.orm import Session
 
 from nowcasting_datamodel.models import (
-    ForecastSQL,
     InputDataLastUpdatedSQL,
     MLModelSQL,
     PVSystemSQL,
     national_gb_label,
 )
-from nowcasting_datamodel.models.forecast import ForecastValueSQL
+from nowcasting_datamodel.models.forecast import ForecastValueSQL, ForecastSQL
 from nowcasting_datamodel.models.gsp import LocationSQL
 from nowcasting_datamodel.read.read import get_location, get_model
