+++ nowcasting_datamodel/models/models.py
@@ -10,15 +10,13 @@ The following class are made
 """
 import logging
 from datetime import datetime
-from typing import List, Optional
+from typing import Optional
 
 from pydantic import Field, validator
-from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Index, Integer, String
+from sqlalchemy import Column, DateTime, Index, Integer, String
 from sqlalchemy.orm import relationship
 
 from nowcasting_datamodel.models.base import Base_Forecast
-from nowcasting_datamodel.models.forecast import ForecastValue
-from nowcasting_datamodel.models.gsp import Location
 from nowcasting_datamodel.models.utils import CreatedMixin, EnhancedBaseModel
 from nowcasting_datamodel.utils import datetime_must_have_timezone
 
@@ -27,10 +25,6 @@ national_gb_label = "National-GB"
 logger = logging.getLogger(__name__)
 # TODO #3 Add forecast latest table, this make it easy to load the latest forecast
 
-
-########
-# 2. Location
-########
 
 
 ########
@@ -112,125 +106,6 @@ class InputDataLastUpdated(EnhancedBaseModel):
             pv=self.pv,
             satellite=self.satellite,
         )
-
-
-########
-# 6. Forecasts
-########
-# TODO add model_name to forecast, or add model table #13
-class ForecastSQL(Base_Forecast, CreatedMixin):
-    """Forecast SQL model"""
-
-    __tablename__ = "forecast"
-
-    id = Column(Integer, primary_key=True)
-    forecast_creation_time = Column(DateTime(timezone=True))
-
-    # Two distinuise between
-    # 1. a Forecast with some forecast values, for that moment,
-    # 2. a Forecast with all the historic latest value.
-    # we use this boolean.
-    # This make it easier to load the forecast showing historic values very easily
-    historic = Column(Boolean, default=False)
-
-    model = relationship("MLModelSQL", back_populates="forecast")
-    model_id = Column(Integer, ForeignKey("model.id"), index=True)
-
-    # many (forecasts) to one (location)
-    location = relationship("LocationSQL", back_populates="forecast")
-    location_id = Column(Integer, ForeignKey("location.id"), index=True)
-
-    # one (forecasts) to many (forecast_value)
-    forecast_values = relationship("ForecastValueSQL", back_populates="forecast")
-    forecast_values_latest = relationship(
-        "ForecastValueLatestSQL", back_populates="forecast_latest"
-    )
-
-    # many (forecasts) to one (input_data_last_updated)
-    input_data_last_updated = relationship("InputDataLastUpdatedSQL", back_populates="forecast")
-    input_data_last_updated_id = Column(
-        Integer, ForeignKey("input_data_last_updated.id"), index=True
-    )
-
-    Index("index_forecast", CreatedMixin.created_utc.desc())
-
-
-class Forecast(EnhancedBaseModel):
-    """A single Forecast"""
-
-    location: Location = Field(..., description="The location object for this forecaster")
-    model: MLModel = Field(..., description="The name of the model that made this forecast")
-    forecast_creation_time: datetime = Field(
-        ..., description="The time when the forecaster was made"
-    )
-    historic: bool = Field(
-        False,
-        description="if False, the forecast is just the latest forecast. "
-        "If True, historic values are also given",
-    )
-    forecast_values: List[ForecastValue] = Field(
-        ...,
-        description="List of forecasted value objects. Each value has the datestamp and a value",
-    )
-    input_data_last_updated: InputDataLastUpdated = Field(
-        ...,
-        description="Information about the input data that was used to create the forecast",
-    )
-
-    _normalize_forecast_creation_time = validator("forecast_creation_time", allow_reuse=True)(
-        datetime_must_have_timezone
-    )
-
-    def to_orm(self) -> ForecastSQL:
-        """Change model to ForecastSQL"""
-        return ForecastSQL(
-            model=self.model.to_orm(),
-            forecast_creation_time=self.forecast_creation_time,
-            location=self.location.to_orm(),
-            input_data_last_updated=self.input_data_last_updated.to_orm(),
-            forecast_values=[forecast_value.to_orm() for forecast_value in self.forecast_values],
-            historic=self.historic,
-        )
-
-    @classmethod
-    def from_orm_latest(cls, forecast_sql: ForecastSQL):
-        """Method to make Forecast object from ForecastSQL,
-
-        but move 'forecast_values_latest' to 'forecast_values'
-        This is useful as we want the API to still present a Forecast object.
-        """
-        # do normal transform
-        forecast = cls.from_orm(forecast_sql)
-
-        # move 'forecast_values_latest' to 'forecast_values'
-        forecast.forecast_values = [
-            ForecastValue.from_orm(forecast_value)
-            for forecast_value in forecast_sql.forecast_values_latest
-        ]
-
-        return forecast
-
-    def normalize(self):
-        """Normalize forecasts by installed capacity mw"""
-        self.forecast_values = [
-            forecast_value.normalize(self.location.installed_capacity_mw)
-            for forecast_value in self.forecast_values
-        ]
-
-        return self
-
-
-class ManyForecasts(EnhancedBaseModel):
-    """Many Forecasts"""
-
-    forecasts: List[Forecast] = Field(
-        ...,
-        description="List of forecasts for different GSPs",
-    )
-
-    def normalize(self):
-        """Normalize forecasts by installed capacity mw"""
-        self.forecasts = [forecast.normalize() for forecast in self.forecasts]
 
 
 ########
