+++ nowcasting_datamodel/read/read.py
@@ -14,8 +14,6 @@ from sqlalchemy.orm.session import Session
 
 from nowcasting_datamodel import N_GSP
 from nowcasting_datamodel.models import (
-    ForecastSQL,
-    ForecastValueSQL,
     InputDataLastUpdatedSQL,
     LocationSQL,
     MLModelSQL,
@@ -23,6 +21,7 @@ from nowcasting_datamodel.models import (
     StatusSQL,
     national_gb_label,
 )
+from nowcasting_datamodel.models.forecast import ForecastValueSQL, ForecastSQL
 from nowcasting_datamodel.models.forecast import ForecastValueLatestSQL
 
 logger = logging.getLogger(__name__)
